Move function stereotype_watches to stereotype module
