+++ gaphor/SysML/requirements/requirement.py
@@ -21,9 +21,8 @@ from gaphor.UML.classes.klass import (
     attributes_compartment,
     operation_watches,
     operations_compartment,
-    stereotype_compartments,
-    stereotype_watches,
 )
+from gaphor.UML.classes.stereotype import stereotype_compartments, stereotype_watches
 from gaphor.UML.recipes import stereotypes_str
